requirements.txt
Django==2.1.7
django-dotenv==1.4.2
django-getenv==1.3.2

djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
django-filter==2.1.0

privex-loghelper>=1.0.2
privex-steemengine==1.0.1
privex-jsonrpc>=1.0.1

mysqlclient==1.4.2.post1
psycopg2==2.7.7 --no-binary psycopg2

requests==2.21.0
beem==0.20.19
python-dateutil==2.8.0
